[Refactor] Divide constants files
