+++ src/js/constants/alert.js
@@ -1,0 +1,5 @@
+export const ALERT = {
+  INVALID_CARNAME:
+    '유효하지 않은 이름 길이입니다. 자동차의 이름은 1자이상, 5자 이하만 가능합니다.',
+  RACING_END: '레이싱이 끝났습니다',
+};
