+++ src/js/constants/elements.js
@@ -1,0 +1,15 @@
+export const ELEMENT = {
+  DATAID: {
+    CAR_NAME_INPUT: 'name-input',
+    MOVE_SUBMIT: 'move-submit',
+    MOVE_INPUT: 'move-input',
+    SUMIT_CARNAME: 'submit-carname',
+  },
+  CLASS: {
+    CAR_NAME_INPUT: 'name-input',
+    MOVE_INPUT: 'move-input',
+    MOVE_SUBMIT_BUTTON: 'move-submit-button',
+    TRIAL_COUNT_WRAPPER: 'trial-count-wrapper',
+    NAME_SUBMIT_BUTTON: 'name-submit-button',
+  },
+};
