+++ src/js/constants/state.js
@@ -1,0 +1,10 @@
+export const DEFAULT_STORE_STATE = {
+  carNames: '',
+  trialNumber: null,
+  isVisibleTrial: false,
+  isVisibleProgress: false,
+  isVisibleResult: false,
+  isRacingEnd: false,
+  racingMap: null,
+  winners: '',
+};
