+++ src/js/constants/validation.js
@@ -1,0 +1,1 @@
+export const MAX_TRIAL_NUMBER = 30;
