--- src/js/index 2.js
@@ -1,3 +1,0 @@
-import { App } from './App.js';
-
-new App({ $target: document.querySelector('#app') });
